update MOCK failure output

+++ source/h2_allocate.hpp
@@ -3,7 +3,7 @@ template <typename T>
 class h2_allocator {
  public:
    typedef size_t size_type;
-   typedef ptrdiff_t difference_type;
+   typedef long long difference_type;
    typedef T* pointer;
    typedef const T* const_pointer;
    typedef T& reference;
